+++ examples/with-dynamic/src/components/normal.tsx
@@ -1,0 +1,7 @@
+export default () => {
+  return <div>normal text</div>;
+};
+
+export function NameExportComp() {
+  return <div>name exported</div>;
+}