+++ examples/with-dynamic/src/pages/normal/basic.tsx
@@ -1,0 +1,9 @@
+import { dynamic } from '@ice/runtime';
+
+const Normal = dynamic(() => import('../../components/normal'), {
+  fallback: () => <div>normal fallback</div>,
+});
+
+export default () => {
+  return <Normal />;
+};