+++ examples/with-dynamic/src/pages/normal/name-export.tsx
@@ -1,0 +1,13 @@
+import { dynamic } from '@ice/runtime';
+
+const Normal = dynamic(
+  import('../../components/normal').then((mod) => {
+    return {
+      default: mod.NameExportComp,
+    };
+  }),
+);
+
+export default () => {
+  return <Normal />;
+};