+++ packages/ice/src/bundler/rspack/getConfig.ts
@@ -1,4 +1,6 @@
+import * as path from 'path';
 import getRspackConfig from '@ice/rspack-config';
+import type { TaskConfig } from 'build-scripts';
 import type { Configuration, rspack as Rspack } from '@rspack/core';
 import type { Config } from '@ice/shared-config/types';
 import { getRouteExportConfig } from '../../service/config.js';
@@ -11,7 +13,6 @@ import {
 } from '../../constant.js';
 import { getReCompilePlugin, getServerPlugin, getSpinnerPlugin } from '../config/plugins.js';
 import { getExpandedEnvs } from '../../utils/runtimeEnv.js';
-import DataLoaderPlugin from '../../webpack/DataLoaderPlugin.js';
 import type { BundlerOptions, Context } from '../types.js';
 
 type GetConfig = (
@@ -33,16 +34,14 @@ const getConfig: GetConfig = async (context, options, rspack) => {
   const {
     rootDir,
     userConfig,
-    getAllPlugin,
     extendsPluginAPI: {
       serverCompileTask,
       getRoutesFile,
-      generator,
     },
   } = context;
   const { reCompile, ensureRoutesConfig } = getRouteExportConfig(rootDir);
   const getPlugins = (taskConfig: Config): Config['plugins'] => {
-    const { target, outputDir, useDataLoader, server } = taskConfig;
+    const { target, outputDir, server } = taskConfig;
     return [
       // Add spinner for webpack task.
       getSpinnerPlugin(spinner),
@@ -60,14 +59,6 @@ const getConfig: GetConfig = async (context, options, rspack) => {
       }),
       // Add ReCompile plugin when routes config changed.
       getReCompilePlugin(reCompile, routeManifest),
-      // Add DataLoader plugin.
-      useDataLoader && new DataLoaderPlugin({
-        serverCompiler,
-        target,
-        rootDir,
-        getAllPlugin,
-        frameworkExports: generator.getExportList('framework', target),
-      }),
     ].filter(Boolean) as Config['plugins'];
   };
   return await Promise.all(taskConfigs.map(async ({ config }) => {
@@ -91,5 +82,51 @@ const getConfig: GetConfig = async (context, options, rspack) => {
   }));
 };
 
+type GetDataLoaderRspackConfig = (
+  context: Context,
+  task: TaskConfig<Config>,
+  rspack: typeof Rspack,
+) => Promise<Configuration>;
+
+export const getDataLoaderConfig: GetDataLoaderRspackConfig = async (context, task, rspack) => {
+  const {
+    rootDir,
+    extendsPluginAPI: {
+      generator,
+    },
+  } = context;
+  const { config } = task;
+  const frameworkExports = generator.getExportList('framework', config.target);
+  return await getRspackConfig({
+    rootDir,
+    rspack,
+    runtimeTmpDir: RUNTIME_TMP_DIR,
+    getExpandedEnvs,
+    runtimeDefineVars: {
+      [IMPORT_META_TARGET]: JSON.stringify(config.target),
+      [IMPORT_META_RENDERER]: JSON.stringify('client'),
+    },
+    localIdentName: config.cssModules?.localIdentName || (config.mode === 'development' ? CSS_MODULES_LOCAL_IDENT_NAME_DEV : CSS_MODULES_LOCAL_IDENT_NAME),
+    taskConfig: {
+      ...config,
+      // Override task config for dataLoader.
+      assetsManifest: false,
+      entry: {
+        'data-loader': path.join(rootDir, RUNTIME_TMP_DIR, 'data-loader.ts'),
+      },
+      swcOptions: {
+        keepExports: ['dataLoader'],
+      },
+      splitChunks: false,
+      redirectImports: frameworkExports,
+      // Data loader should be hot reload in development mode.
+      fastRefresh: false,
+      devServer: {
+        hot: false,
+      },
+      name: 'dataLoader',
+    },
+  });
+};
 
 export default getConfig;