+++ packages/ice/src/constant.ts
@@ -67,16 +67,22 @@ export const RUNTIME_EXPORTS = [
       'useSuspenseData',
       'usePublicAppContext',
       'Await',
-      'defineDataLoader',
-      'defineServerDataLoader',
-      'defineStaticDataLoader',
       'usePageLifecycle',
       'unstable_useDocumentData',
+      'dynamic',
     ],
     alias: {
       usePublicAppContext: 'useAppContext',
     },
     source: '@ice/runtime',
+  },
+  {
+    specifier: [
+      'defineDataLoader',
+      'defineServerDataLoader',
+      'defineStaticDataLoader',
+    ],
+    source: '@ice/runtime/data-loader',
   },
 ];