+++ packages/ice/src/createService.ts
@@ -57,6 +57,7 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
   const coreTemplate = path.join(templateDir, 'core/');
   const configFile = commandArgs.config || 'ice.config.(mts|mjs|ts|js|cjs|json)';
   const dataCache = new Map<string, string>();
+  const builtinPlugin = commandArgs.plugin as string;
   const generator = new Generator({
     // Directory of templates includes `core` and `exports`.
     templateDir,
@@ -161,7 +162,16 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
   if (target === WEB) {
     plugins.push(pluginWeb());
   }
-
+  if (builtinPlugin) {
+    try {
+      const pluginModule = await dynamicImport(builtinPlugin.startsWith('.') ? path.join(rootDir, builtinPlugin) : builtinPlugin);
+      const plugin = pluginModule.default || pluginModule;
+      plugins.push(plugin());
+    } catch (err) {
+      logger.error(`Load builtin plugin error, Faild to import plugin "${builtinPlugin}".`);
+      throw err;
+    }
+  }
   // Register framework level API.
   RUNTIME_EXPORTS.forEach(exports => {
     generatorAPI.addExport(exports);
@@ -285,6 +295,7 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
   dataCache.set('routes', JSON.stringify(routesInfo));
   dataCache.set('hasExportAppData', hasExportAppData ? 'true' : '');
 
+  const hasDataLoader = Boolean(userConfig.dataLoader) && (hasExportAppData || Boolean(routesInfo.loaders));
   // Render exports files if route component export dataLoader / pageConfig.
   renderExportsTemplate(
     {
@@ -386,6 +397,7 @@ async function createService({ rootDir, command, commandArgs }: CreateServiceOpt
         },
         userConfig,
         configFile,
+        hasDataLoader,
       };
       try {
         if (command === 'test') {